Revert "#17935: yolov10x model demo bringup (#20255)"
This reverts commit 809d8b90317ec430d0a9a3106bfd63291f3074f5.
