+++ .github/CODEOWNERS
@@ -201,7 +201,7 @@ models/demos/utils/llm_demo_utils.py @skhorasganiTT @mtairum @uaydonat
 models/experimental/stable_diffusion_xl_base @mbezuljTT @pavlepopovic @ipotkonjak-tt @tenstorrent/metalium-developers-convolutions
 models/demos/mobilenetv2 @tenstorrent/cse-developer-ttnn
 models/experimental/yolov8s_world @tenstorrent/cse-developer-ttnn
-models/demos/yolov10 @tenstorrent/cse-developer-ttnn
+models/experimental/yolov10 @tenstorrent/cse-developer-ttnn
 models/demos/yolov9c @tenstorrent/cse-developer-ttnn
 models/demos/yolov8x @tenstorrent/cse-developer-ttnn
