+++ models/experimental/yolov10/reference/yolov10.py
@@ -2,10 +2,9 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import copy
-
 import torch
 import torch.nn as nn
+import copy
 
 
 def dist2bbox(distance, anchor_points, xywh=True, dim=-1):
