+++ models/experimental/yolov10/tt/attention.py
@@ -2,10 +2,9 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import math
-
 import ttnn
-from models.demos.yolov10.tt.common import Conv
+import math
+from models.experimental.yolov10.tt.common import Conv
 
 
 class TtnnAttention:
