+++ models/experimental/yolov10/tt/bottleneck.py
@@ -3,11 +3,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.demos.yolov10.tt.common import Conv
+from models.experimental.yolov10.tt.common import Conv
 
 
-class TtnnSCDown:
-    def __init__(self, device=None, parameters=None, conv_pt=None, auto_shard=False):
+class TtnnBottleNeck:
+    def __init__(self, shortcut=True, device=None, parameters=None, conv_pt=None):
+        self.shortcut = shortcut
         self.device = device
         self.parameters = parameters
         self.conv_pt = conv_pt
@@ -22,14 +23,12 @@ class TtnnSCDown:
             device,
             parameters.cv2,
             self.conv_pt.cv2,
-            enable_identity=True,
-            use_1d_systolic_array=False,
-            auto_shard=auto_shard,
             deallocate_activation=True,
         )
 
     def __call__(self, input_tensor):
         cv1 = self.cv1(input_tensor)
-        cv1 = ttnn.sharded_to_interleaved(cv1, ttnn.L1_MEMORY_CONFIG)
-        output = self.cv2(cv1)
-        return output
+        cv2 = self.cv2(cv1)
+        if input_tensor.get_layout() == ttnn.ROW_MAJOR_LAYOUT:
+            input_tensor = ttnn.to_layout(input_tensor, ttnn.TILE_LAYOUT)
+        return ttnn.add(input_tensor, cv2, memory_config=ttnn.L1_MEMORY_CONFIG) if self.shortcut else cv2
