+++ models/experimental/yolov10/tt/c2f.py
@@ -3,8 +3,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.demos.yolov10.tt.bottleneck import TtnnBottleNeck
-from models.demos.yolov10.tt.common import Conv, deallocate_tensors
+from models.experimental.yolov10.tt.bottleneck import TtnnBottleNeck
+from models.experimental.yolov10.tt.common import Conv, deallocate_tensors
 
 
 class TtnnC2f:
