+++ models/experimental/yolov10/tt/c2fcib.py
@@ -3,9 +3,9 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.demos.yolov10.tt.cib import TtnnCIB
-from models.demos.yolov10.tt.common import Conv, deallocate_tensors
+from models.experimental.yolov10.tt.common import Conv, deallocate_tensors
 from models.experimental.yolo_common.yolo_utils import concat
+from models.experimental.yolov10.tt.cib import TtnnCIB
 
 
 class TtnnC2fCIB:
