+++ models/experimental/yolov10/tt/common.py
@@ -2,9 +2,8 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import math
-
 import ttnn
+import math
 from models.experimental.yolo_common.yolo_utils import determine_num_cores, get_core_grid_from_num_cores
 
 
@@ -22,7 +21,7 @@ def interleaved_to_sharded(x):
     return ttnn.reshard(x, shardspec) if x.is_sharded() else ttnn.interleaved_to_sharded(x, shardspec)
 
 
-class TtYolov10_Conv2D:
+class TtYolov10Conv2D:
     def __init__(
         self,
         conv,
@@ -153,7 +152,7 @@ class Conv:
         self.enable_act = enable_act
         if not self.enable_identity:
             activation = "silu"
-        self.conv = TtYolov10_Conv2D(
+        self.conv = TtYolov10Conv2D(
             parameter.conv,
             conv_pt.conv,
             device=device,
