+++ models/experimental/yolov10/tt/model_preprocessing.py
@@ -2,12 +2,11 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
+import ttnn
 import torch
 import torch.nn as nn
-from ttnn.model_preprocessing import fold_batch_norm2d_into_conv2d, infer_ttnn_module_args, preprocess_model_parameters
-
-import ttnn
-from models.demos.yolov10.reference.yolov10 import Conv, YOLOv10
+from ttnn.model_preprocessing import preprocess_model_parameters, fold_batch_norm2d_into_conv2d, infer_ttnn_module_args
+from models.experimental.yolov10.reference.yolov10 import YOLOv10, Conv
 
 
 def make_anchors(device, feats, strides, grid_cell_offset=0.5):
