+++ models/experimental/yolov10/tt/scdown.py
@@ -2,13 +2,12 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
+from models.experimental.yolov10.tt.common import Conv
 import ttnn
-from models.demos.yolov10.tt.common import Conv
 
 
-class TtnnBottleNeck:
-    def __init__(self, shortcut=True, device=None, parameters=None, conv_pt=None):
-        self.shortcut = shortcut
+class TtnnSCDown:
+    def __init__(self, device=None, parameters=None, conv_pt=None, auto_shard=False):
         self.device = device
         self.parameters = parameters
         self.conv_pt = conv_pt
@@ -23,12 +22,16 @@ class TtnnBottleNeck:
             device,
             parameters.cv2,
             self.conv_pt.cv2,
+            enable_identity=True,
+            use_1d_systolic_array=False,
+            auto_shard=auto_shard,
             deallocate_activation=True,
         )
 
     def __call__(self, input_tensor):
         cv1 = self.cv1(input_tensor)
-        cv2 = self.cv2(cv1)
-        if input_tensor.get_layout() == ttnn.ROW_MAJOR_LAYOUT:
-            input_tensor = ttnn.to_layout(input_tensor, ttnn.TILE_LAYOUT)
-        return ttnn.add(input_tensor, cv2, memory_config=ttnn.L1_MEMORY_CONFIG) if self.shortcut else cv2
+        cv1 = ttnn.sharded_to_interleaved(
+            cv1, ttnn.L1_MEMORY_CONFIG
+        )  # needed since cv2 uses block_sharding and input is in height sharding
+        output = self.cv2(cv1)
+        return output
