+++ models/experimental/yolov10/tt/sppf.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.demos.yolov10.tt.common import Conv, deallocate_tensors
+from models.experimental.yolov10.tt.common import Conv, deallocate_tensors
 from models.experimental.yolo_common.yolo_utils import concat
